Login problem solvedgit status

+++ src/main/java/br/com/dio/REST_spring/domain/entity/product/ProductRequestDTO.java
@@ -10,5 +10,6 @@ public record ProductRequestDTO(
         String name,
 
         @NotNull
-        Double price
-) {}
+        Integer price
+) {
+}
