+++ src/main/java/br/com/dio/REST_spring/domain/entity/product/ProductResponseDTO.java
@@ -1,7 +1,8 @@
 package br.com.dio.REST_spring.domain.entity.product;
 
-public record ProductResponseDTO(String id, String name, Double price) {
-    public ProductResponseDTO(Product product) {
+public record ProductResponseDTO(String id, String name, Integer price) {
+    public ProductResponseDTO(Product product){
         this(product.getId(), product.getName(), product.getPrice());
     }
 }
+
